test: Test methods of Solana WebSocket RPC. (2) - Investigated and documented errors in Solana's official `BlockSubscribe` example. - Refactored `Transaction` decoding logic to align with updated Solana specs. - Improved compatibility with Moshi adapters for custom decoding requirements.

+++ src/main/java/net/deanly/solana/sdk/types/UiTokenAmount.java
@@ -22,7 +22,7 @@ public class UiTokenAmount {
     /// Token amount as a float, accounting for decimals. DEPRECATED
     @Deprecated
     @Json(name = "uiAmount")
-    private Long uiAmount;
+    private Float uiAmount;
 
     /// Token amount as a string, accounting for decimals.
     @Json(name = "uiAmountString")
